+++ utils/vrad/vraddetailprops.cpp
@@ -361,7 +361,7 @@ public:
 	CLightSurface(int iThread) : m_pSurface(0), m_HitFrac(1.0f), m_bHasLuxel(false), m_iThread(iThread) {}
 
 	// call back with a node and a context
-	bool EnumerateNode( int node, Ray_t const& ray, float f, int context )
+	bool EnumerateNode( int node, Ray_t const& ray, float f, intp context )
 	{
 		dface_t* pSkySurface = 0;
 
@@ -410,7 +410,7 @@ public:
 	}
 
 	// call back with a leaf and a context
-	virtual bool EnumerateLeaf( int leaf, Ray_t const& ray, float start, float end, int context )
+	virtual bool EnumerateLeaf( int leaf, Ray_t const& ray, float start, float end, intp context )
 	{
 		bool hit = false;
 		dleaf_t* pLeaf = &dleafs[leaf];
